+++ toegeungil/src/public/community/components/CommunityDetails.js
@@ -3,12 +3,15 @@ import { useParams, Link } from "react-router-dom";
 // import CommunityLocation from "./CommunityLocation";
 import DetailsStyle from './css/CommunityDetails.module.css';
 import UserNickName from "./UserNickName";
+import jwt_decode from "jwt-decode";
+
 
 
 const CommunityDetails = () => {
     const { communityNum } = useParams();
     const [community, setCommunitys] = useState([]);
     const [userNum, setUserNum] = useState(null);
+    const [user, setUser] = useState();
 
     const getCommunitys = () => {
         fetch(process.env.REACT_APP_URL + `/communitys/${communityNum}`)
@@ -19,14 +22,18 @@ const CommunityDetails = () => {
             });
     };
 
-    useEffect(() => {},[community])
+    useEffect(() => {
+        if (sessionStorage.getItem("Authorizaton")) {
+            setUser(jwt_decode(sessionStorage.getItem("Authorizaton")))
+        }
+    }, [community]);
 
     const deleteApi = (communityNum) => {
         console.log(communityNum)
-        if(window.confirm("커뮤니티글을 삭제하시겠습니까?")){
-            fetch(process.env.REACT_APP_URL + `/communitys/${communityNum}`,{
+        if (window.confirm("커뮤니티글을 삭제하시겠습니까?")) {
+            fetch(process.env.REACT_APP_URL + `/communitys/${communityNum}`, {
                 method: "DELETE",
-                headers:{
+                headers: {
 
                 },
             }).then(res => res.json()).then(res => {
@@ -54,12 +61,17 @@ const CommunityDetails = () => {
                     </div>
                 </div>
             </div>
-            <Link to="/communitys/modify" state={{ "communitys": community }}>
-                <button className={DetailsStyle.CommunityModifyButton}>수정</button>
-            </Link>
-            <Link to="/communitys">
-            <button className={DetailsStyle.CommunityDeleteButton} onClick={() => deleteApi(community.communityNum)}>삭제</button>
-            </Link>
+            
+            {(user && user.no === community.userNum) ? (
+                <>
+                    <Link to="/communitys/modify" state={{ "communitys": community }}>
+                        <button className={DetailsStyle.CommunityModifyButton}>수정</button>
+                    </Link>
+                    <Link to="/communitys">
+                        <button className={DetailsStyle.CommunityDeleteButton} onClick={() => deleteApi(community.communityNum)}>삭제</button>
+                    </Link>
+                </>
+            ) : null}
         </>
     )
