+++ toegeungil/src/public/community/components/CommunityList.js
@@ -5,9 +5,11 @@ import '../components/css/CommunityMain.css';
 import { Link } from "react-router-dom";
 import UserNickName from "./UserNickName";
 import DetailsTitleStyle from "../components/css/CommunityDetailsTitle.module.css";
+import jwt_decode from "jwt-decode"
 
 const CommunityList = () => {
     const [communityList, setCommunityList] = useState([]);
+    const [user, setUser] = useState();
 
     const getCommunityList = () => {
         fetch(process.env.REACT_APP_URL + "/communitys")
@@ -20,9 +22,14 @@ const CommunityList = () => {
     };
 
     useEffect(() => {
+        if (sessionStorage.getItem("Authorizaton")) {
+            setUser(jwt_decode(sessionStorage.getItem("Authorizaton")))
+        }
+
         getCommunityList();
         console.log(communityList);
     }, []);
+
 
     const viewCommunity = (communityNum) => {
         window.location.href = `/communitys/${communityNum}`;
@@ -68,9 +75,10 @@ const CommunityList = () => {
                         ))}
                 </tbody>
             </table>
+            {!user ? null : (!user?.auth[0] == "USER" ||  !user?.auth[0] == "TUTOR" || !user?.auth[0] == "ADMIN" ) ? "회원이 아닙니다." : 
             <Link to={"/communitys/write"}>
                 <button className="community-regist-button" >커뮤니티 글 작성</button>
-            </Link>
+            </Link>}
         </div>
     );
 }
