+++ toegeungil/src/public/community/pages/CommunityWrite.js
@@ -2,7 +2,8 @@ import CommunityPosting from "../components/CommunityPosting";
 import layout from "../../layout/layout.css";
 
 
-function CommunityWrite({}) {
+
+function CommunityWrite({ }) {
     return (
         <>
             <div className={layout.layout}>
